+++ CHANGELOG.md
@@ -6,7 +6,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [0.16.0] - WIP
 ## Modified
-  * Adjust `UnitQuaternion`s and `Rotation3`s generated from the `Standard` distribution to be uniformly distributed.
+  * Adjust `UnitQuaternion`s, `Rotation3`s, and `Rotation2`s generated from the `Standard` distribution to be uniformly
+    distributed.
 ### Added
   * Add construction of a `Point` from an array by implementing the `From` trait.
   * Add support for generating uniformly distributed random unit column vectors using the `Standard` distribution.